requirements.txt
amqp==5.0.6
arrow==1.1.0
asgiref==3.3.4
attrs==21.2.0
billiard==3.6.4.0
blessed==1.18.1
celery==5.1.0
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
Django==3.2.4
django-filter==1.1.0
django-model-utils==4.1.1
django-picklefield==3.0.1
django-q==1.3.8
django-rest-swagger==2.2.0
django-smtp-ssl==1.0
django-tinymce==3.3.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
drf-yasg==1.20.0
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jsonschema==3.2.0
kiwisolver==1.3.1
kombu==5.1.0
MarkupSafe==2.0.1
matplotlib==3.4.2
method==0.0.0.0.dev20200926
numpy==1.20.3
openapi-codec==1.3.2
packaging==20.9
Pillow==8.2.0
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
PyJWT==2.1.0
pyotp==2.6.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
ruamel.yaml==0.17.9
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.16.0
sqlparse==0.4.1
swagger-spec-validator==2.7.3
translation==1.0.5
uritemplate==3.0.1
urllib3==1.26.5
utils==1.0.1
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.2.0
